Clarify confusing code in fci/rdm.py
linkstr and linkstrl have the same number of columns, and the call
to direct_spin1._unpack produces the latter, but rdm creation needs
the former. I had confusingly used direct_spin1._unpack to build
the default linkstrl strictly for the purposes of checking the number
of columns, and then passed the unmodified kwarg "link_index" to the
rdm build functions (which have their own, correct, defaults). This
created an incorrect impression of the necessary dimensions of
link_index in my_pyscf/fci/rdm.py if you didn't notice that the
thing I called "linkstr" and the thing I called "link_index" were
not the same thing. Now, I use the correct "unpack" function in
my_pyscf/fci/rdm.py, and to be sure it's correct I pass its output
to the rdm build functions.